Squashed commit of the following:
commit b1d2a9256e109012ec5724b97eceea282719c596
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Thu Mar 14 16:00:33 2024 +0100

    feat(multicompte): MB-2074 , manage delete add component add account on multicompte screen

commit 43d7d339763a86d9521ce2b836cfa60df147fec7
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Thu Mar 14 15:57:24 2024 +0100

    fix: add-account was broken from account-selection screen

commit 1aa3e6201926f08024c1572c6e9058410a1aa1bc
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Thu Mar 14 15:33:03 2024 +0100

    update(stack): last stack update before final rc build

commit 6cb617d348a90c6ab1dcdbcb579a47c5a15a55a9
Merge: 50a655055 301dfce42
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Thu Mar 14 15:22:51 2024 +0100

    Merge branch 'dev/1.12.0' of https://github.com/opendigitaleducation/edifice-mobile-framework into dev/1.12.0

commit 50a655055ba2ed7e4133b0cd541c345400d6bc82
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Thu Mar 14 15:22:40 2024 +0100

    feat(multicompte): MB-2074 , manage delete all account from multicompte screen

commit 301dfce42c8167b8805fe3db50f6f2ab54720842
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Thu Mar 14 15:09:00 2024 +0100

    chore: erase user preferences when account is deleted

commit aea7cb9ef86f8cecd83ba448df5f7e701331a027
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Thu Mar 14 12:16:03 2024 +0100

    feat(multicompte): MB-2074 , add remove account action

commit 38c82ec70832a8cc3888502ddb3578ae3def2dd2
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Wed Mar 13 18:14:29 2024 +0100

    fix(auth): add account now clears previous session data

commit 5216dd773f396590ccdbee5ea2ce7e75a772a9c4
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Wed Mar 13 18:02:36 2024 +0100

    fix(auth): use previous account token to remove fcm token

commit 74b3184f4143f59ec94081d7233b1dcbec3a96b9
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Wed Mar 13 17:46:03 2024 +0100

    fix(auth): switching accounts now clears previous session data

commit 93a8c76cb4976c9f9b0e0399292443ea3d8dbeb9
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Wed Mar 13 14:44:32 2024 +0100

    fix(i18n): MB-2566, Change user-page-editlang EN/ES translations

commit 20faccfc1e2d4c2598d4182c3dfade053609cc37
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Wed Mar 13 13:54:32 2024 +0100

    fix: make storage setPrefix idempotent

commit b5a51540b61d3d7cdc1e27714f852647057cb141
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 18:02:59 2024 +0100

    chore(auth): add missing i18n

commit b7bd0f3146a48901e001ffb069a7bc1f3cf37b72
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 17:34:55 2024 +0100

    chore(competences): MB-2016 new storage for competences

commit e76b8391ef2d32ec7499d663f05604ba11e3167c
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Tue Mar 12 17:07:45 2024 +0100

    feat(multicompte): MB-2025 , switch account from user screen

commit b9a99ba4c7df21b6117ff713d0920a4b6b64092d
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 15:48:38 2024 +0100

    chore(myapps): rename module-config

commit 6191a1984062db6878d5d90edfcce6047040bf88
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 15:48:38 2024 +0100

    chore: storageName is now mandatory in module-config

commit ad0312059324c8b0a36ac8d136358a1198fb6e51
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 15:48:38 2024 +0100

    chore(timeline): MB-2016 new storage for myapps

commit 9eb1e329fa3b7737fc5c7faf9334b4aa7c6d2909
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 15:48:38 2024 +0100

    chore(storage): rename 'sessionStorage' to 'preferences' + type checks

commit 288366f31440583431f2b7f83fcc243d01b62357
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Tue Mar 12 15:21:10 2024 +0100

    chore(build): Fix Nokogiri version

commit 378055e19cd2cc17184ae9a7f152965274b70326
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Tue Mar 12 14:21:48 2024 +0100

    chore(timeline): MB-2016 new storage for timeline

commit b3c8aa59e068a32b1db1583631624fd0bd1cd8f7
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Tue Mar 12 12:27:33 2024 +0100

    chore(i18n): Missing translations addition

commit 55517a4fe1f3f77d1a6075909ca52d559c4dd282
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 19:58:36 2024 +0100

    chore: add i18n keys + remove unused components

commit 81aa1e4ac63a9499fe28718f380c6f489d8c6ae3
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 18:24:59 2024 +0100

    chore: oversimplify new storage api :D (part 3)

commit a43211bf804224a728457f0cd878b9aa89e9e227
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 17:46:44 2024 +0100

    chore: oversimplify new storage api :D (part 2)

commit 9b4ce74be92dde58e98487c024a2ba92a529397b
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 17:42:20 2024 +0100

    chore: oversimplify new storage api :D

commit 9783da6b09d4437db3b6634624f873a41a633f03
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Mon Mar 11 16:42:41 2024 +0100

    chore(i18n): Missing translations addition

commit 149a3b5a82100412a8857390ef633aec2f2308fb
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Mon Mar 11 16:13:22 2024 +0100

    update(stack): Update React Native to 0.73.6

commit 35f749bd6fa073a8878c7b197cc974936cf1684e
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Mon Mar 11 13:16:57 2024 +0100

    release: 1.12.0-rc.1(1120013)

commit d0f2fdf1f74f396317a658955a378c962df1b862
Author: Guillaume DROULEZ <guillaume.droulez@edifice.io>
Date:   Mon Mar 11 12:19:25 2024 +0100

    update(stack): Update some packages.

commit 5dceef7cce97335ca8cfc4ab48fe909550de3e92
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 12:29:49 2024 +0100

    fix(auth): MB-2569 account migration was broken

commit 895130db0483d56fa151dc63da37161e20a7d687
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 11:34:59 2024 +0100

    chore: add getOrderedAccounts getter

commit fad8bf3e5cfd7aa204fd4ddf31fab0be74310ebc
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Mon Mar 11 11:32:25 2024 +0100

    fix(auth): MB-2568 preserve accounts order

commit 753c60dcf44c4cfe67030a46625f78fe556ae214
Author: Valentin Mourot <valentin.mourot@opendigitaleducation.com>
Date:   Fri Mar 8 19:05:19 2024 +0100

    release: 1.12.0-alpha.12(1120012)

commit 6df5ac38977c54151597dbf22d1aa2686e605796
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Fri Mar 8 18:33:34 2024 +0100

    fix(auth): fix handle-account-list

commit 7e28d28ed3f9127432c8809637a574715e71e86e
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Fri Mar 8 17:46:14 2024 +0100

    chore(avatar): add error handling

commit c25bd892033341e25ef089b21fe51bd4d4d11f6f
Author: Valentin Mourot <thaledric@gmail.com>
Date:   Fri Mar 8 17:38:19 2024 +0100

    chore: display loader on account-select screen

commit 35e902309ebb8eebbc9de7dd02cad953c0553b57
Author: Léa DE AMORIN <leadeamorin.ps@gmail.com>
Date:   Fri Mar 8 16:33:12 2024 +0100

    chore(wayf): MB-2537, duplicate loginWayf & wayf screens

+++ src/framework/modules/auth/actions.ts
@@ -4,7 +4,14 @@ import { ThunkAction, ThunkDispatch } from 'redux-thunk';
 
 import { I18n } from '~/app/i18n';
 import { IGlobalState } from '~/app/store';
-import { ERASE_ALL_ACCOUNTS, IAuthState, actions, assertSession, getState as getAuthState } from '~/framework/modules/auth/reducer';
+import {
+  ERASE_ALL_ACCOUNTS,
+  IAuthState,
+  actions,
+  assertSession,
+  getState as getAuthState,
+  getSession,
+} from '~/framework/modules/auth/reducer';
 import appConf, { Platform } from '~/framework/util/appConf';
 import { Error } from '~/framework/util/error';
 import { createEndSessionAction } from '~/framework/util/redux/reducerFactory';
@@ -20,6 +27,7 @@ import {
   AuthLoggedAccount,
   AuthPendingRedirection,
   AuthRequirement,
+  AuthSavedAccount,
   AuthSavedAccountWithTokens,
   ForgotMode,
   IActivationError,
@@ -55,6 +63,7 @@ import {
   readSavedStartup,
   readShowOnbording,
   writeCreateAccount,
+  writeDeleteAccount,
   writeLogout,
   writeReplaceAccount,
 } from './storage';
@@ -275,15 +284,18 @@ export const loginSteps = {
 
 const requirementsThatNeedLegalUrls = [AuthRequirement.MUST_REVALIDATE_TERMS, AuthRequirement.MUST_VALIDATE_TERMS];
 
-export function deactivateLoggedAccountAction(action?: AnyAction) {
+export function deactivateLoggedAccountActionIfApplicable(action?: AnyAction) {
   return async (dispatch: ThunkDispatch<any, any, any>, getState: () => any) => {
-    const account = assertSession();
-    // Unregister the device token from the backend
-    await authService.removeFirebaseTokenWithAccount(account);
-    // Erase requests cache
-    await clearRequestsCacheLegacy();
-    // flush sessionReducers
-    dispatch(createEndSessionAction());
+    const account = getSession();
+    if (account) {
+      // Unregister the device token from the backend
+      await authService.removeFirebaseTokenWithAccount(account);
+      // Erase requests cache
+      await clearRequestsCacheLegacy();
+      // flush sessionReducers
+      dispatch(createEndSessionAction());
+      dispatch(actions.deactivate());
+    }
     if (action) dispatch(action);
   };
 }
@@ -327,12 +339,12 @@ const getLoginFunctions = {
       success:
         (...args: Parameters<typeof actions.addAccount>) =>
         async (dispatch: AuthDispatch, getState: () => IGlobalState) => {
-          await dispatch(deactivateLoggedAccountAction(actions.addAccount(...args)));
+          await dispatch(deactivateLoggedAccountActionIfApplicable(actions.addAccount(...args)));
         },
       requirement:
         (...args: Parameters<typeof actions.addAccountRequirement>) =>
         async (dispatch: AuthDispatch, getState: () => IGlobalState) => {
-          await dispatch(deactivateLoggedAccountAction(actions.addAccountRequirement(...args)));
+          await dispatch(deactivateLoggedAccountActionIfApplicable(actions.addAccountRequirement(...args)));
         },
       activation: actions.addAccountActivation,
       passwordRenew: actions.addAccountPasswordRenew,
@@ -353,12 +365,12 @@ const getLoginFunctions = {
       success:
         (...args: Parameters<typeof actions.addAccount>) =>
         async (dispatch: AuthDispatch, getState: () => IGlobalState) => {
-          await dispatch(deactivateLoggedAccountAction(actions.replaceAccount(id, ...args)));
+          await dispatch(deactivateLoggedAccountActionIfApplicable(actions.replaceAccount(id, ...args)));
         },
       requirement:
         (...args: Parameters<typeof actions.addAccountRequirement>) =>
         async (dispatch: AuthDispatch, getState: () => IGlobalState) => {
-          await dispatch(deactivateLoggedAccountAction(actions.replaceAccountRequirement(id, ...args)));
+          await dispatch(deactivateLoggedAccountActionIfApplicable(actions.replaceAccountRequirement(id, ...args)));
         },
       activation: actions.redirectActivation,
       passwordRenew: actions.redirectPasswordRenew,
@@ -684,6 +696,16 @@ export function quietLogoutAction() {
 export function manualLogoutAction() {
   return async (dispatch: ThunkDispatch<any, any, any>, getState: () => any) => {
     await dispatch(quietLogoutAction());
+  };
+}
+
+export function removeAccountAction(account: AuthLoggedAccount | AuthSavedAccountWithTokens | AuthSavedAccount) {
+  return async (dispatch: ThunkDispatch<any, any, any>, getState: () => any) => {
+    if (accountIsLogged(account)) {
+      await dispatch(quietLogoutAction());
+    }
+    dispatch(actions.removeAccount(account.user.id));
+    writeDeleteAccount(account.user.id);
   };
 }
